fixed issue with incorrect station durations on run once
Trying to apply "run once" to a program resulted in erroneous
scheduling of station durations.  The root cause was the modification
of the contents of the ethernet response buffer returned by
get_buffer().  The buffer was being modified because as part of
resetting all stations, another web call is made to record valve
status in an external influx db.
The fix is to use a bit more memory by duplicating the buffer for
local use, freeing it upon return.